pool: add `pool::internal` module
